Enum uni_gl::BlendMode

pub enum BlendMode {

    Zero = 0,
    One = 1,
    SrcColor = 768,
    OneMinusSrcColor = 769,
    SrcAlpha = 770,
    OneMinusSrcAlpha = 771,
    DstAlpha = 772,
    OneMinusDstAlpha = 773,
    DstColor = 774,
    OneMinusDstColor = 775,
    SrcAlphaSaturate = 776,
    ConstantColor = 32_769,
    OneMinusConstantColor = 32_770,
    ConstantAlpha = 32_771,
    OneMinusConstantAlpha = 32_772,

}

Constants passed to WebGLRenderingContext.blendFunc() or WebGLRenderingContext.blendFuncSeparate() to specify the blending mode (for both, RBG and alpha, or separately).

Variants

Zero = 0

Passed to blendFunc or blendFuncSeparate to turn off a component.

One = 1

Passed to blendFunc or blendFuncSeparate to turn on a component.

SrcColor = 768

Passed to blendFunc or blendFuncSeparate to multiply a component by the source elements color.

OneMinusSrcColor = 769

Passed to blendFunc or blendFuncSeparate to multiply a component by one minus the source elements color.

SrcAlpha = 770

Passed to blendFunc or blendFuncSeparate to multiply a component by the source’s alpha.

OneMinusSrcAlpha = 771

Passed to blendFunc or blendFuncSeparate to multiply a component by one minus the source’s alpha.

DstAlpha = 772

Passed to blendFunc or blendFuncSeparate to multiply a component by the destination’s alpha.

OneMinusDstAlpha = 773

Passed to blendFunc or blendFuncSeparate to multiply a component by one minus the destination’s alpha.

DstColor = 774

Passed to blendFunc or blendFuncSeparate to multiply a component by the destination’s color.

OneMinusDstColor = 775

Passed to blendFunc or blendFuncSeparate to multiply a component by one minus the destination’s color.

SrcAlphaSaturate = 776

Passed to blendFunc or blendFuncSeparate to multiply a component by the minimum of source’s alpha or one minus the destination’s alpha.

ConstantColor = 32_769

Passed to blendFunc or blendFuncSeparate to specify a constant color blend function.

OneMinusConstantColor = 32_770

Passed to blendFunc or blendFuncSeparate to specify one minus a constant color blend function.

ConstantAlpha = 32_771

Passed to blendFunc or blendFuncSeparate to specify a constant alpha blend function.

OneMinusConstantAlpha = 32_772

Passed to blendFunc or blendFuncSeparate to specify one minus a constant alpha blend function.