Enum uni_gl::Parameter

pub enum Parameter {

    BlendEquation,
    BlendEquationAlpha,
    BlendDstRgb,
    BlendSrcRgb,
    BlendDstAlpha,
    BlendSrcAlpha,
    BlendColor,
    ArrayBufferBinding,
    ElementArrayBufferBinding,
    LineWidth,
    AliasedPointSizeRange,
    AliasedLineWidthRange,
    CullFaceMode,
    FrontFace,
    DepthRange,
    DepthWritemask,
    DepthClearValue,
    DepthFunc,
    StencilClearValue,
    StencilFunc,
    StencilFail,
    StencilPassDepthFail,
    StencilPassDepthPass,
    StencilRef,
    StencilValueMask,
    StencilWritemask,
    StencilBackFunc,
    StencilBackFail,
    StencilBackPassDepthFail,
    StencilBackPassDepthPass,
    StencilBackRef,
    StencilBackValueMask,
    StencilBackWritemask,
    Viewport,
    ScissorBox,
    ColorClearValue,
    ColorWritemask,
    UnpackAlignment,
    PackAlignment,
    MaxTextureSize,
    MaxViewportDims,
    SubpixelBits,
    RedBits,
    GreenBits,
    BlueBits,
    AlphaBits,
    DepthBits,
    StencilBits,
    PolygonOffsetUnits,
    PolygonOffsetFactor,
    TextureBinding2d,
    SampleBuffers,
    Samples,
    SampleCoverageValue,
    SampleCoverageInvert,
    CompressedTextureFormats,
    Vendor,
    Renderer,
    Version,
    ImplementationColorReadType,
    ImplementationColorReadFormat,
    BrowserDefaultWebgl,
    TextureBindingCubeMap,
    MaxCubeMapTextureSize,

}

Constants passed to WebGLRenderingContext.getParameter() to specify what information to return.

Variants

BlendEquation

Passed to getParameter to get the current RGB blend function. same as BlendEquationRgb

BlendEquationAlpha

Passed to getParameter to get the current alpha blend function. Same as BLEND_EQUATION

BlendDstRgb

Passed to getParameter to get the current destination RGB blend function.

BlendSrcRgb

Passed to getParameter to get the current destination RGB blend function.

BlendDstAlpha

Passed to getParameter to get the current destination alpha blend function.

BlendSrcAlpha

Passed to getParameter to get the current source alpha blend function.

BlendColor

Passed to getParameter to return a the current blend color.

ArrayBufferBinding

Passed to getParameter to get the array buffer binding.

ElementArrayBufferBinding

Passed to getParameter to get the current element array buffer.

LineWidth

Passed to getParameter to get the current lineWidth (set by the lineWidth method).

AliasedPointSizeRange

Passed to getParameter to get the current size of a point drawn with gl.POINTS

AliasedLineWidthRange

Passed to getParameter to get the range of available widths for a line. Returns a length-2 array with the lo value at 0, and hight at 1.

CullFaceMode

Passed to getParameter to get the current value of cullFace. Should return FRONT, BACK, or FRONT_AND_BACK

FrontFace

Passed to getParameter to determine the current value of frontFace. Should return CW or CCW.

DepthRange

Passed to getParameter to return a length-2 array of floats giving the current depth range.

DepthWritemask

Passed to getParameter to determine if the depth write mask is enabled.

DepthClearValue

Passed to getParameter to determine the current depth clear value.

DepthFunc

Passed to getParameter to get the current depth function. Returns NEVER, ALWAYS, LESS, EQUAL, LEQUAL, GREATER, GEQUAL, or NOTEQUAL.

StencilClearValue

Passed to getParameter to get the value the stencil will be cleared to.

StencilFunc

Passed to getParameter to get the current stencil function. Returns NEVER, ALWAYS, LESS, EQUAL, LEQUAL, GREATER, GEQUAL, or NOTEQUAL.

StencilFail

Passed to getParameter to get the current stencil fail function. Should return KEEP, REPLACE, INCR, DECR, INVERT, INCR_WRAP, or DECR_WRAP.

StencilPassDepthFail

Passed to getParameter to get the current stencil fail function should the depth buffer test fail. Should return KEEP, REPLACE, INCR, DECR, INVERT, INCR_WRAP, or DECR_WRAP.

StencilPassDepthPass

Passed to getParameter to get the current stencil fail function should the depth buffer test pass. Should return KEEP, REPLACE, INCR, DECR, INVERT, INCR_WRAP, or DECR_WRAP.

StencilRef

Passed to getParameter to get the reference value used for stencil tests.

StencilValueMask

StencilWritemask

StencilBackFunc

StencilBackFail

StencilBackPassDepthFail

StencilBackPassDepthPass

StencilBackRef

StencilBackValueMask

StencilBackWritemask

Viewport

Returns an Int32Array with four elements for the current viewport dimensions.

ScissorBox

Returns an Int32Array with four elements for the current scissor box dimensions.

ColorClearValue

ColorWritemask

UnpackAlignment

PackAlignment

MaxTextureSize

MaxViewportDims

SubpixelBits

RedBits

GreenBits

BlueBits

AlphaBits

DepthBits

StencilBits

PolygonOffsetUnits

PolygonOffsetFactor

TextureBinding2d

SampleBuffers

Samples

SampleCoverageValue

SampleCoverageInvert

CompressedTextureFormats

Vendor

Renderer

Version

ImplementationColorReadType

ImplementationColorReadFormat

BrowserDefaultWebgl

TextureBindingCubeMap

MaxCubeMapTextureSize